Split sitemap out to its own controller and moved routing out of the controllers
